api/requirements.txt
django-cors-headers==3.7.0
django-environ==0.4.5